MenuActivity uses ViewModel to prevent splash from showing at every screen rotation
