+++ app/src/main/AndroidManifest.xml
@@ -11,32 +11,29 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
-
-        </activity>
-        <activity
-            android:name=".SplashScreen"
-            android:theme="@style/AppTheme">
+        <activity android:name=".MenuActivity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".TouchActivity">
+        <activity android:name=".SurveyActivity.TouchActivity">
 
         </activity>
-        <activity android:name=".toiletFeedback.FragmentLayout">
+        <activity android:name=".SurveyActivity.FragmentLayout"
+            android:configChanges="orientation|screenLayout">
             <meta-data
                 android:name="RECOGNIZE_MODE"
                 android:value="1" />
         </activity>
-        <activity android:name=".toiletFeedback.SubFragmentLayout">
+        <activity android:name=".SurveyActivity.SubFragmentLayout"
+            android:configChanges="orientation|screenLayout">
             <meta-data
                 android:name="RECOGNIZE_MODE"
                 android:value="1" />
         </activity>
 
-        <service android:name=".toiletFeedback.Services.SpeechService" />
+        <service android:name=".SurveyActivity.Services.SpeechService" />
     </application>
