+++ app/src/main/java/com/anewtech/phone/client/MenuActivity/MenuViewModel.java
@@ -1,0 +1,12 @@
+package com.anewtech.phone.client.MenuActivity;
+
+import android.arch.lifecycle.ViewModel;
+
+/**
+ * Created by heriz on 29/12/2017.
+ */
+
+public class MenuViewModel extends ViewModel {
+    //This prevents splash from showing every time the screen rotate
+    boolean initialSplash = true;
+}
