+++ app/src/main/java/com/anewtech/phone/client/SurveyActivity/Fragments/FragmentToActivityInterface.java
@@ -1,0 +1,11 @@
+package com.anewtech.phone.client.SurveyActivity.Fragments;
+
+import com.anewtech.phone.client.SurveyActivity.Models.services.Answernaire;
+
+/**
+ * Created by heriz on 26/12/2017.
+ */
+
+public interface FragmentToActivityInterface {
+    public void OnItemPickedInFragment(int position, Answernaire subAnswer);
+}
