+++ app/src/main/java/com/anewtech/phone/client/SurveyActivity/Models/data/Answer.java
@@ -1,0 +1,13 @@
+package com.anewtech.phone.client.SurveyActivity.Models.data;
+
+/**
+ * Created by heriz on 26/12/2017.
+ */
+
+public class Answer {
+    public String id;
+    public String message;
+    public boolean sub;
+    public String questionid;
+    public String seq;
+}
