+++ app/src/main/java/com/anewtech/phone/client/SurveyActivity/Models/data/FinalMessage.java
@@ -1,0 +1,14 @@
+package com.anewtech.phone.client.SurveyActivity.Models.data;
+
+/**
+ * Created by heriz on 26/12/2017.
+ */
+
+public class FinalMessage {
+    public String id;
+    public String message;
+    public String assigneQuestionOrAnswer;
+    public String answerid;
+    public String seq;
+    public boolean oneforall;
+}
