+++ app/src/main/java/com/anewtech/phone/client/SurveyActivity/Models/data/Question.java
@@ -1,0 +1,14 @@
+package com.anewtech.phone.client.SurveyActivity.Models.data;
+
+/**
+ * Created by heriz on 26/12/2017.
+ */
+
+public class Question {
+    public String id;
+    public String message;
+    public boolean multichoice;
+    public String templateid;
+    public String answerid;
+    public Integer seq;
+}
