+++ app/src/main/java/com/anewtech/phone/client/SurveyActivity/Models/data/Template.java
@@ -1,0 +1,14 @@
+package com.anewtech.phone.client.SurveyActivity.Models.data;
+
+/**
+ * Created by heriz on 26/12/2017.
+ */
+
+public class Template {
+    public String id;
+    public String title;
+    public boolean firstlayer;
+    public boolean secondlayer;
+    public boolean thirdlayer;
+    public Integer type;
+}
