--- app/src/main/java/com/anewtech/phone/client/toiletFeedback/Fragments/FragmentToActivityInterface.java
@@ -1,11 +1,0 @@
-package com.anewtech.phone.client.toiletFeedback.Fragments;
-
-import com.anewtech.phone.client.toiletFeedback.Models.services.Answernaire;
-
-/**
- * Created by heriz on 26/12/2017.
- */
-
-public interface FragmentToActivityInterface {
-    public void OnItemPickedInFragment(int position, Answernaire subAnswer);
-}
