--- app/src/main/java/com/anewtech/phone/client/toiletFeedback/Models/data/Answer.java
@@ -1,15 +1,0 @@
-package com.anewtech.phone.client.toiletFeedback.Models.data;
-
-import android.arch.lifecycle.ViewModel;
-
-/**
- * Created by heriz on 26/12/2017.
- */
-
-public class Answer {
-    public String id;
-    public String message;
-    public boolean sub;
-    public String questionid;
-    public String seq;
-}
