--- app/src/main/java/com/anewtech/phone/client/toiletFeedback/Models/data/FinalMessage.java
@@ -1,16 +1,0 @@
-package com.anewtech.phone.client.toiletFeedback.Models.data;
-
-import android.arch.lifecycle.ViewModel;
-
-/**
- * Created by heriz on 26/12/2017.
- */
-
-public class FinalMessage {
-    public String id;
-    public String message;
-    public String assigneQuestionOrAnswer;
-    public String answerid;
-    public String seq;
-    public boolean oneforall;
-}
