--- app/src/main/java/com/anewtech/phone/client/toiletFeedback/Models/data/Question.java
@@ -1,16 +1,0 @@
-package com.anewtech.phone.client.toiletFeedback.Models.data;
-
-import android.arch.lifecycle.ViewModel;
-
-/**
- * Created by heriz on 26/12/2017.
- */
-
-public class Question {
-    public String id;
-    public String message;
-    public boolean multichoice;
-    public String templateid;
-    public String answerid;
-    public Integer seq;
-}
