--- app/src/main/java/com/anewtech/phone/client/toiletFeedback/Models/data/SurveyDataModel.java
@@ -1,21 +1,0 @@
-package com.anewtech.phone.client.toiletFeedback.Models.data;
-
-import android.arch.lifecycle.ViewModel;
-
-import java.util.List;
-
-/**
- * Created by heriz on 26/12/2017.
- */
-
-public class SurveyDataModel {
-    public String version;
-    public String title;
-    public String datetime;
-    public String uploadedby;
-    public String access;
-    public Template template;
-    public List<Question> layersquestions;
-    public List<Answer> layersanswers;
-    public List<FinalMessage> layerslastmessages;
-}
