--- app/src/main/java/com/anewtech/phone/client/toiletFeedback/Models/data/SurveyViewModel.java
@@ -1,13 +1,0 @@
-package com.anewtech.phone.client.toiletFeedback.Models.data;
-
-import android.arch.lifecycle.ViewModel;
-
-/**
- * Created by heriz on 28/12/2017.
- */
-
-public class SurveyViewModel extends ViewModel {
-    public String answer = "";
-    public String question = "";
-    public int qnNo = 0;
-}
