--- app/src/main/java/com/anewtech/phone/client/toiletFeedback/Models/data/Template.java
@@ -1,14 +1,0 @@
-package com.anewtech.phone.client.toiletFeedback.Models.data;
-
-/**
- * Created by heriz on 26/12/2017.
- */
-
-public class Template {
-    public String id;
-    public String title;
-    public boolean firstlayer;
-    public boolean secondlayer;
-    public boolean thirdlayer;
-    public Integer type;
-}
